[FIX] rangeTokenize: avoid enrichedToken when not necessary
A profiling on a big spreadsheet (732k+ formua cells) showed that out of
28 seconds to start the `Model`, we were spending more than 23 seconds in
the method `normalize`.
After investigation, it turns out we were enriching the tokens and
deenriching them right after just to use a method that did not
specifically need enriched tokens. This commit removes those useless
casts.

After this commit, the time spent in normalize went down from 23 seconds
to 20 seconds. The whole loading time was improved by about 10%.
